Title: 컴퓨터와 함께하는 숫자 게임 2, Time: -1ms, Memory: -1MB, Status: Runtime Error - CodeTree

+++ 240601/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -6,24 +6,24 @@ public class Main {
     public static void main(String[] args) throws Exception {
         BufferedReader reader = new BufferedReader(new InputStreamReader(System.in));
 
-        int startNumber = 1; // 범위의 시작
-        int endNumber = Integer.parseInt(reader.readLine()); // 컴퓨터가 선택한 수 (범위의 끝)
+        long startNumber = 1; // 범위의 시작
+        long endNumber = Integer.parseInt(reader.readLine()); // 컴퓨터가 선택한 수 (범위의 끝)
         
         StringTokenizer tokenizer = new StringTokenizer(reader.readLine());
-        int startTarget = Integer.parseInt(tokenizer.nextToken()); // 목표 숫자 범위의 시작
-        int endTarget = Integer.parseInt(tokenizer.nextToken()); // 목표 숫자 범위의 끝
+        long startTarget = Long.parseLong(tokenizer.nextToken()); // 목표 숫자 범위의 시작
+        long endTarget = Long.parseLong(tokenizer.nextToken()); // 목표 숫자 범위의 끝
         
-        int min = Integer.MAX_VALUE; // 게임이 가장 적게 지속될 때
-        int max = 0; // 게임이 가장 오래 지속될 때
+        long min = Integer.MAX_VALUE; // 게임이 가장 적게 지속될 때
+        long max = 0; // 게임이 가장 오래 지속될 때
         
-        for (int target = startTarget; target <= endTarget; target++) {
-            int left = startNumber; // 구간의 시작 숫자
-            int right = endNumber; // 구간의 끝 숫자
+        for (long target = startTarget; target <= endTarget; target++) {
+            long left = startNumber; // 구간의 시작 숫자
+            long right = endNumber; // 구간의 끝 숫자
             int count = 0;
             
             while (left <= right) {
                 count++;
-                int mid = (left + right) / 2; // 구간의 중앙값
+                long mid = (left + right) / 2; // 구간의 중앙값
                 
                 if (mid == target) {
                     break;
